requirements.txt
asgiref==3.5.2
black==22.6.0
boto3==1.24.56
botocore==1.27.56
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorclass==2.2.2
coverage==6.4.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.5
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
django-import-export==2.8.0
django-storages==1.13.1
djangorestframework==3.13.1
docker==6.0.0
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
isort==5.10.1
jmespath==1.0.1
MarkupPy==1.14
mypy-extensions==0.4.3
numpy==1.23.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
pip-upgrader==1.4.15
pipenv==2022.8.19
pipreqs==0.4.11
platformdirs==2.5.2
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sql==2022.4.0
sqlparse==0.4.2
tablib==3.2.1
terminaltables==3.1.10
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
virtualenv==20.16.3
virtualenv-clone==0.5.7
websocket-client==1.3.3
whitenoise==6.2.0
wsl==0.12
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
